requirements.txt
# AGI Layer v3.9 - CPU-only зависимости

# Основные зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
asyncpg==0.29.0

# Telegram интеграция
python-telegram-bot==20.7

# Машинное обучение (CPU-only)
torch==2.1.1+cpu
torchvision==0.16.1+cpu
torchaudio==2.1.1+cpu
--index-url https://download.pytorch.org/whl/cpu

# Transformers и модели
transformers==4.36.0
diffusers==0.24.0
accelerate==0.25.0
sentence-transformers==2.2.2

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
easyocr==1.7.0

# Векторная база данных
chromadb==0.4.18

# Web UI
streamlit==1.28.1
plotly==5.17.0

# Работа с данными
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2

# Логирование и мониторинг
structlog==23.2.0
prometheus-client==0.19.0

# Утилиты
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
aiofiles==23.2.1

# Разработка и тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Дополнительные зависимости для стабильности
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4

